refactor: auto-split test_perturbation_base.py
